refactor: organize swagger docs
