+++ src/main/docs/components.ts
@@ -1,4 +1,4 @@
-import { apiKeyAuthSchema } from '@/main/docs/schemas/account'
+import { apiKeyAuthSchema } from '@/main/docs/schemas/account/authentication'
 import { badRequest, unauthorized, forbidden, notFound, conflict, serverError } from '@/main/docs/components/http'
 
 export default {
