+++ src/main/docs/paths/account/authentication/index.ts
@@ -1,3 +1,2 @@
-export * from './change-email-path'
 export * from './sign-in-path'
 export * from './sign-up-path'
