+++ src/main/docs/paths/account/authentication/sign-in-path.ts
