+++ src/main/docs/paths/account/authentication/sign-up-path.ts
