+++ src/main/docs/paths/account/commands/change-email-path.ts
