+++ src/main/docs/paths/account/commands/index.ts
@@ -1,0 +1,1 @@
+export * from './change-email-path'
