+++ src/main/docs/paths/product/commands/add-product-path.ts
