+++ src/main/docs/paths/product/commands/index.ts
@@ -1,0 +1,2 @@
+export * from './add-product-path'
+export * from './update-product-path'
