+++ src/main/docs/paths/product/commands/update-product-path.ts
