+++ src/main/docs/paths/product/queries/get-latest-products-path.ts
