+++ src/main/docs/paths/product/queries/get-product-by-slug-path.ts
