+++ src/main/docs/paths/product/queries/get-products-by-category-path.ts
