+++ src/main/docs/paths/product/queries/get-products-with-discount-path.ts
