+++ src/main/docs/paths/product/queries/index.ts
@@ -1,6 +1,4 @@
-export * from './add-product-path'
 export * from './get-latest-products-path'
 export * from './get-product-by-slug-path'
 export * from './get-products-by-category-path'
 export * from './get-products-with-discount-path'
-export * from './update-product-path'
