+++ src/main/docs/schemas.ts
@@ -1,8 +1,14 @@
 import { errorSchema } from '@/main/docs/schemas/error-schema'
 import { successSchema } from '@/main/docs/schemas/success-schema'
-import { accountSchema, signUpInputSchema, signInInputSchema, changeEmailInputSchema } from '@/main/docs/schemas/account'
-import { categorySchema, categoriesSchema, addCategoryInputSchema } from '@/main/docs/schemas/category'
-import { productSchema, productsSchema, productInputSchema } from '@/main/docs/schemas/product'
+
+import { accountSchema, signUpInputSchema, signInInputSchema } from '@/main/docs/schemas/account/authentication'
+import { changeEmailInputSchema } from '@/main/docs/schemas/account/commands'
+
+import { categorySchema, categoriesSchema } from '@/main/docs/schemas/category/queries'
+import { addCategoryInputSchema } from '@/main/docs/schemas/category/commands'
+
+import { productSchema, productsSchema } from '@/main/docs/schemas/product/queries'
+import { productInputSchema } from '@/main/docs/schemas/product/commands'
 
 export default {
   error: errorSchema,
