+++ src/main/docs/schemas/account/authentication/account-schema.ts
