+++ src/main/docs/schemas/account/authentication/api-key-auth-schema.ts
