+++ src/main/docs/schemas/account/authentication/index.ts
@@ -1,5 +1,4 @@
 export * from './account-schema'
 export * from './api-key-auth-schema'
-export * from './change-email-input-schema'
 export * from './sign-in-input-schema'
 export * from './sign-up-input-schema'
