+++ src/main/docs/schemas/account/authentication/sign-up-input-schema.ts
