+++ src/main/docs/schemas/account/commands/change-email-input-schema.ts
