+++ src/main/docs/schemas/account/commands/index.ts
@@ -1,0 +1,1 @@
+export * from './change-email-input-schema'
