+++ src/main/docs/schemas/category/commands/add-category-input-schema.ts
