+++ src/main/docs/schemas/category/commands/index.ts
@@ -1,0 +1,1 @@
+export * from './add-category-input-schema'
