+++ src/main/docs/schemas/category/queries/categories-schema.ts
