+++ src/main/docs/schemas/category/queries/category-schema.ts
