+++ src/main/docs/schemas/category/queries/index.ts
@@ -1,3 +1,2 @@
-export * from './add-category-input-schema'
 export * from './categories-schema'
 export * from './category-schema'
