+++ src/main/docs/schemas/product/commands/index.ts
@@ -1,0 +1,1 @@
+export * from './product-input-schema'
