+++ src/main/docs/schemas/product/commands/product-input-schema.ts
