+++ src/main/docs/schemas/product/queries/index.ts
@@ -1,3 +1,2 @@
-export * from './product-input-schema'
 export * from './product-schema'
 export * from './products-schema'
