+++ src/main/docs/schemas/product/queries/product-schema.ts
