+++ src/main/docs/schemas/product/queries/products-schema.ts
